% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abacus.R
\name{abacus.detail}
\alias{abacus.detail}
\title{A more detailed abacus plot}
\usage{
abacus.detail(
  X,
  plot.title = "Tag detections over time",
  x.breaks = "7 days",
  DT.limits,
  x.exp = c(0, 0),
  y.exp = c(0, 0)
)
}
\arguments{
\item{X}{a list of "hit" data.frames, each element is a different individual.}

\item{plot.title}{Title}

\item{x.breaks}{Character string of "X days" to set tick marks/gridlines.}

\item{DT.limits}{min and max dates to plot, as a vector of character strings}

\item{x.exp}{x-axis expand() values for making plots pretty}

\item{y.exp}{y-axis expand() values for making plots pretty}
}
\value{
A plot where the individual tower detections are now slightly offset.
}
\description{
Simple plot has symbols for different towers overwriting each other. This
assumes you will split your data and plot a subset of individuals at a time.
It then offsets individual tower detections for the same bird slightly, so
detections by multiple towers at approximately the same time are discernible.
}
