% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abacus.R
\name{abacus.simple}
\alias{abacus.simple}
\title{Simple abacus plots for detection data}
\usage{
abacus.simple(
  X,
  plot.title = "Tag detections over time",
  x.breaks = "7 days",
  DT.limits
)
}
\arguments{
\item{X}{data.frame of hits}

\item{plot.title}{title}

\item{x.breaks}{Character string of "X days" to set tick marks/gridlines.}

\item{DT.limits}{min and max dates to plot, as a vector of character strings}
}
\value{
a plot with each tag on one line, for all unique tags (mfgID field)
a different color and symbol for each tower
}
\description{
Plot each tag one one line, different color/symbol. Set up for nuthatch data
only right now, based on tower names/ number
}
