% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsim.R
\name{bhnu.pop.sim}
\alias{bhnu.pop.sim}
\title{Stochastic simulation of female-only abundance. Single annual breeding
attempt}
\usage{
bhnu.pop.sim(
  N.sim,
  N.female,
  N.years,
  ext.thresh = 0,
  surv.params,
  breed.params,
  nest.params,
  fledge.params
)
}
\arguments{
\item{N.sim}{Number of simulations}

\item{N.female}{Number of females initially}

\item{N.years}{Number of years to project}

\item{ext.thresh}{(Quasi) extinction threshold abundance}

\item{surv.params}{1 value if no stochasticity in survival; 2 parameters
mean/sd}

\item{breed.params}{1 value if no stochasticity in breeding probability; 2
parameters mean/sd}

\item{nest.params}{1 value if no stochasticity in nest survival probability;
2 parameters mean/sd}

\item{fledge.params}{2 parameters to simulate the number of fledglings per
successful nest. Generated via a binomial with the first parameter is the
max number of fledglings that can be produced (N), and the second is the
probability (p_. N*p analogus to lambda from Poisson distribution.}
}
\value{
Two element list of the projection matrix Nsim x N.years, and the
probability of extinction.
}
\description{
Stochastic simulation of female-only abundance. Single annual breeding
attempt
}
