% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{close_motus}
\alias{close_motus}
\title{Close motus database connection}
\usage{
close_motus(my.conn)
}
\arguments{
\item{my.conn}{The connection object returned by open.motus.}
}
\value{
closes the database connection
}
\description{
Close motus database connection
}
