% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsim.R
\name{fxn.1year}
\alias{fxn.1year}
\title{Project N for one year to the nex}
\usage{
fxn.1year(
  phi.f = 0.5,
  p.breed = 1,
  p.fledge = c(5, 0.6),
  p.success = 0.65,
  N = 50
)
}
\arguments{
\item{phi.f}{survival probability}

\item{p.breed}{breeding probability}

\item{p.fledge}{2 parameter describing the number of fledglings produced as a binomial}

\item{p.success}{the probability a nest survives to produce fledglings}

\item{N}{the intial abundance}
}
\description{
Project N for one year to the nex
}
