% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{open.motus}
\alias{open.motus}
\title{A  wrapper to connect to a motus database}
\usage{
\method{open}{motus}(path)
}
\arguments{
\item{path}{The full path to where the .motus RSQlite database is located}
}
\value{
A database connection
}
\description{
A  wrapper to connect to a motus database
}
